chore: minor enhancements
- Flat struture
- Uses goreleaser changelog
- Updates golangci-lint
