+++ .travis.yml
@@ -7,9 +7,7 @@ env:
 
 before_install:
   # Install linters and misspell
-  # FIXME: Since version 1.12.4 issue with mal formatted file that contains dependency with -
-  # https://github.com/golangci/golangci-lint/issues/138#issuecomment-449931610
-  - curl -sfL https://install.goreleaser.com/github.com/golangci/golangci-lint.sh | bash -s -- -b $GOPATH/bin v1.12.3
+  - curl -sfL https://install.goreleaser.com/github.com/golangci/golangci-lint.sh | bash -s -- -b $GOPATH/bin v1.14.0
   - golangci-lint --version
 
 install:
@@ -20,16 +18,6 @@ deploy:
   - provider: script
     skip_cleanup: true
     script: curl -sL https://git.io/goreleaser | bash
-    on:
-      tags: true
-      condition: $TRAVIS_GO_VERSION =~ ^1\.x$
-
-  - provider: releases
-    api_key: ${GITHUB_TOKEN}
-    file: dist/run-semaphoreci-build_*
-    skip_cleanup: true
-    overwrite: true
-    file_glob: true
     on:
       tags: true
       condition: $TRAVIS_GO_VERSION =~ ^1\.x$
