+++ go.mod
@@ -2,8 +2,8 @@ module github.com/mmatur/run-semaphoreci-build
 
 require (
 	github.com/containous/flaeg v1.4.1
-	github.com/davecgh/go-spew v1.1.1
-	github.com/ldez/go-semaphoreci v1.2.1
+	github.com/ldez/go-semaphoreci v1.2.2
 	github.com/ogier/pflag v0.0.1
 	github.com/pkg/errors v0.8.1 // indirect
+	gotest.tools v2.3.0+incompatible // indirect
 )
